microservices/payment/requirements-lock.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
entrypoints==0.3
flake8==3.7.9
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
mccabe==0.6.1
paypalrestsdk==1.13.1
protobuf==3.11.3
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pyOpenSSL==19.1.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
